% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_check.R
\name{anova_check}
\alias{anova_check}
\title{anova_check}
\usage{
anova_check(model)
}
\arguments{
\item{model}{A linear model generated via "glm", "lm" or "aov".}
}
\value{
Graphic object
}
\description{
Returns diagnostics graphs for glm, lm or aov object.
}
\examples{
anova_check(lm(bill_length_mm ~ bill_depth_mm, data = penguins))
anova_check(lm_object)
anova_check(anova_object) # Error
}
