% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{cooks}
\alias{cooks}
\title{cooks}
\usage{
cooks(
  fitted.lm,
  label = TRUE,
  show.threshold = FALSE,
  threshold = "convention",
  scale.factor = 0.5
)
}
\arguments{
\item{fitted.lm}{A linear model generated via "glm", "lm" or "aov".}

\item{label}{Logical. Whether or not outliers should be labeled on the plot.}

\item{show.threshold}{Logical. Whether or not the threshold value should be on the plot.}

\item{threshold}{Determines the formula used in the calculation of the threshold value,
used to determine whether an observation is an outlier. "baseR" (0.5 and 1), "matlab" (mean(cooksd)*3), and "convention" (4/n and 1).}

\item{scale.factor}{Integer value that scales the point size and linewidth to allow customized viewing. Defaults to 0.5}
}
\value{
gg object
}
\description{
Returns Cook's distance plot for supplied linear models. The code for this function is based on the "gg_cooksd" function from the "lindia" package.
}
\examples{
model <- lm(bill_length_mm ~ bill_depth_mm + year, penguins)
cooks(model)
}
\references{
\insertRef{lindia}{classpackage}
}
