% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qq_plots.R
\name{dependent_qq_plot}
\alias{dependent_qq_plot}
\title{dependent_qq_plot}
\usage{
dependent_qq_plot(data, variable, grouping_variable, first_group, second_group)
}
\arguments{
\item{data}{Data in either long or wide format.}

\item{variable}{The variable for which the QQ plot is being constructed.
For wide datasets, this alters the titles of the QQ plot.}

\item{grouping_variable}{The variable that contains the grouping indicator for data in long format.
For data in wide format, this only alters the titles of the QQ plot.}

\item{first_group}{The variable from which the other is subtracted. For data in
long format, this is the level of the grouping variable we would like to be first.
For data in wide format, this is the column's name containing the first set of measurements.}

\item{second_group}{Analogous to first_group.}
}
\value{
Graphic object
}
\description{
Constructs QQ plots for dependent t-test.
}
\examples{
dependent_qq_plot(a1c_measurements, "time_point", "first_measurement", "second_measurement")
}
