% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_a_name.R
\name{draw_a_name}
\alias{draw_a_name}
\title{Draw a name}
\usage{
draw_a_name(pren = "Vincent", s = "M")
}
\arguments{
\item{pren}{The name you want to plot - default is Vincent}

\item{s}{The genre of the name you want to plot useful for mixt names - default is M}
}
\value{
a ggplot of the historical trend of the name
}
\description{
Draw a name
}
\examples{
draw_a_name()
draw_a_name("Louis","M")
draw_a_name("Dominique","F")

}
