% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Draw_names.R
\name{draw_names}
\alias{draw_names}
\title{Draw a name}
\usage{
draw_names(pren = c("Vincent", "Louis"))
}
\arguments{
\item{pren}{The names you want to plot - default is Vincent and Louis}
}
\value{
a ggplot of the historical trend of the two names
}
\description{
a function that plots the historical trend of two names
}
\examples{
draw_names()
draw_names("Dominique","Louis")

}
