% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qq_plots.R
\name{independent_qq_plot}
\alias{independent_qq_plot}
\title{independent_qq_plot}
\usage{
independent_qq_plot(data, variable, grouping_variable, ...)
}
\arguments{
\item{data}{Data in either long or wide format.}

\item{variable}{The variable for which the QQ plot is being constructed.
For wide datasets, this alters the titles of the QQ plot.}

\item{grouping_variable}{The variable that contains the grouping indicator for data in long format.
For data in wide format, this only alters the titles of the QQ plot.}

\item{...}{For data in wide format, a character vector containing the column names of the measurements must be supplied.}
}
\value{
Graphic object
}
\description{
Constructs QQ plots for independet t-tests
}
\examples{
independent_qq_plot(penguins, "bill_length_mm", "sex")
}
