% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_csv_reader.R
\name{my_csv_reader}
\alias{my_csv_reader}
\title{My_csv reader}
\usage{
my_csv_reader(folder = getwd())
}
\arguments{
\item{folder}{the directory where the csv files are saved (default to the working directory)}
}
\value{
a list of dataframes
}
\description{
A function that reads all the csv files saved in a directory and outputs them as a list of dataframes
}
\examples{
my_csv_reader()
my_csv_reader("/Users/XXX/Documents/MyCsvs")

}
