% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multi_excel.R
\name{my_excel_reader}
\alias{my_excel_reader}
\title{Multi_excel}
\usage{
my_excel_reader(path)
}
\arguments{
\item{path}{the path of the excel file}
}
\value{
a list of dataframes
}
\description{
Reads a whole excel file (xlsx) at once and saves every sheet as a dataframe in a list
}
\examples{
my_csv_reader("/Users/XXX/Documents/excel.xlsx")
}
