% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normality_and_correlation.R
\name{normality_correlation}
\alias{normality_correlation}
\title{normality_correlation}
\usage{
normality_correlation(data, method, digits_desired = NULL)
}
\arguments{
\item{data}{Data with numeric variables.}

\item{method}{The method used to calculate correlation:
c("pearson", "kendall", "spearman")}

\item{digits_desired}{Optional argument for desired number of digits in
numeric output. Default value is 5.}
}
\value{
List containing qq plots, correlation matrix, and Shapiro-Wilk test
results
}
\description{
Allows you to check the normality of each numeric variable by conducting the
Shapiro-Wilk test and examining QQ plots. Additionally, it provides
a correlation matrix based on the specified method.
}
\examples{
normality_correlation(penguins, "pearson")
mutate_all(mtcars, as.factor) \%>\% normality_correlation(., "kendall") # Error
normality_correlation(mtcars, "invalid_method") # Error
}
