% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qq_plots.R
\name{one_qq_plot}
\alias{one_qq_plot}
\title{one_qq_plot}
\usage{
one_qq_plot(data, variable)
}
\arguments{
\item{data}{Data in either long or wide format.}

\item{variable}{The variable for which the QQ plot is being constructed.}
}
\value{
Graphic object
}
\description{
Constructs QQ plots for single vectors.
}
\examples{
one_qq_plot(penguins, "bill_length_mm")
}
