% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Save_as_csv.R
\name{save_as_csv}
\alias{save_as_csv}
\title{Save_as_csv}
\usage{
save_as_csv(x, name = "test.csv")
}
\arguments{
\item{x}{a data frame}

\item{name}{an output name}
}
\value{
the path of where the file was saved
}
\description{
A function that saves a dataframe as a .csv file to the working directory and outputs the path of the file (silentlty)
}
\examples{
 head <- head(prenoms) #avoid writing 3M lines
 save_as_csv(head,"prenoms.csv")

}
