% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PackageDoc.R
\docType{package}
\name{SebastianSchwarzPackage-package}
\alias{SebastianSchwarzPackage-package}
\alias{SebastianSchwarzPackage}
\title{A summary package for the R class}
\description{
Utility containing 5 functions we wrote in classe, those being
save_as_csv: simple csv write function
my_csv_reader: simple csv read function
multi_excel: excel import function
draw_names: plot function on French demographical name data
draw_a_name: simple plot function on French demographical name data
}
\author{
Sebastian Schwarz <sebastian.schwarz@polytechnique.edu>
}
