% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superior_almost_sas.R
\name{superior_almost_sas}
\alias{superior_almost_sas}
\title{superior_almost_sas}
\usage{
superior_almost_sas(model, ...)
}
\arguments{
\item{model}{A linear model generated via "lm" or "aov".}

\item{...}{Optional argument for supplying binwidth argument to geom_histogram.}
}
\value{
Graphic object
}
\description{
Works similarly to almost_sas. Returns diagnostics graphs for lm or aov object.
}
\examples{
superior_almost_sas(lm(bill_length_mm ~ bill_depth_mm, data = penguins))
superior_almost_sas(lm_object, binwidth = 10)
superior_almost_sas(anova_object) # Error
}
