% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{outlier_count}
\alias{outlier_count}
\title{outlier_count}
\usage{
outlier_count(df, model)
}
\arguments{
\item{df}{dataset of interest.}

\item{model}{A linear model generated via "glm", "lm" or "aov".}
}
\description{
Returns a tibble with count of suspected outliers in the dataset.
}
\examples{
model <- lm(bill_length_mm ~ bill_depth_mm + year, penguins)
outlier_count(penguins, model)
}
