% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{outlier_graph}
\alias{outlier_graph}
\title{outlier_grpah}
\usage{
outlier_graph(df, model, x_var, y_var, x_lab = NULL, y_lab = NULL)
}
\arguments{
\item{df}{dataset of interest.}

\item{model}{A linear model generated via "glm", "lm" or "aov".}

\item{x_var}{The variable to be plotted on the x-axis.}

\item{y_var}{The variable to be plotted on the y-axis.}

\item{x_lab}{String to be used as the x-axis label. If NULL, defaults to the name of the x_var.}

\item{y_lab}{String to be used as the y-axis label. If NULL, defaults to the name of the y_var.}
}
\description{
Returns a scatterplot augmented with outlier information.
}
\examples{
model <- lm(bill_length_mm ~ bill_depth_mm + year, penguins)
outlier_graph(penguins, model, x_var = bill_depth_mm, y_var = bill_length_mm) # Axis labels will be bill_depth_mm and bill_length_mm, respectively.
outlier_graph(penguins, model, x_var = bill_depth_mm, y_var = bill_length_mm, x_lab = "a", y_lab = "b") # Axis labels will be a and b, respectively.
}
